refactor: 코드 개선
