+++ cypress/e2e/lotto.cy.js
@@ -1,5 +1,5 @@
-import { LOTTO_NUMBER } from '../../src/js/service/Constant';
-import { generateLottoNumbers } from '../../src/js/service/LottoBuyer';
+import { LOTTO_LENGTH, LOTTO_NUMBER } from '../../src/js/service/Constant';
+import { generateLottoNumbers } from '../../src/js/service/lottoShop';
 
 describe('로또 난수 생성 로직을 점검한다', () => {
   it('로또 난수가 최소~최대 범위에 맞게 잘 생성되는지 검사한다', () => {
@@ -14,12 +14,12 @@ describe('로또 난수 생성 로직을 점검한다', () => {
   it('로또가 6자리 생성되는지 검사한다', () => {
     const lottoNumbers = generateLottoNumbers();
 
-    expect(lottoNumbers.length).to.eq(LOTTO_NUMBER.LENGTH);
+    expect(lottoNumbers.length).to.eq(LOTTO_LENGTH);
   });
 
   it('로또 난수가 중복없이 생성되는지 검사한다', () => {
     const lottoNumbers = generateLottoNumbers();
     const deduplicatedLottoNumbers = Array.from(new Set(lottoNumbers));
-    expect(deduplicatedLottoNumbers.length).to.eq(LOTTO_NUMBER.LENGTH);
+    expect(deduplicatedLottoNumbers.length).to.eq(LOTTO_LENGTH);
   });
 });
