+++ src/js/view/ManualInput.js
@@ -1,4 +1,4 @@
-import { LOTTO_NUMBER, LOTTO_PRICE } from '../service/Constant.js';
+import { LOTTO_LENGTH, LOTTO_NUMBER, LOTTO_PRICE } from '../service/Constant.js';
 import { MANUAL_NUMBER_INPUT } from './Element.js';
 import { onLottoNumberInput } from './Listener.js';
 import { $manualNumberList } from './Selector.js';
@@ -21,7 +21,7 @@ export const getPriceOfBuyingManually = (manualNumberItemLength) => {
 };
 
 export const appendLottoNumberInputs = () => {
-  const inputs = Array.from({ length: LOTTO_NUMBER.LENGTH }).map(
+  const inputs = Array.from({ length: LOTTO_LENGTH }).map(
     () => `<input type="text" class="manual-number mx-1 mb-3 text-center" required />`
   );
   const button = `<button type="button" class="btn btn-pink w-10 mx-1">삭제</button>`;
